decouple Frostbite and Hypothermia
