+++ src/main/java/org/terasology/climateConditions/HypothermiaSystem.java
@@ -15,46 +15,23 @@
  */
 package org.terasology.climateConditions;
 
-import org.terasology.alterationEffects.speed.StunAlterationEffect;
-import org.terasology.audio.StaticSound;
-import org.terasology.audio.events.PlaySoundEvent;
-import org.terasology.context.Context;
 import org.terasology.entitySystem.entity.EntityRef;
-import org.terasology.entitySystem.entity.lifecycleEvents.OnAddedComponent;
 import org.terasology.entitySystem.event.ReceiveEvent;
-import org.terasology.entitySystem.prefab.Prefab;
-import org.terasology.entitySystem.prefab.PrefabManager;
 import org.terasology.entitySystem.systems.BaseComponentSystem;
 import org.terasology.entitySystem.systems.RegisterMode;
 import org.terasology.entitySystem.systems.RegisterSystem;
-import org.terasology.logic.characters.*;
-import org.terasology.logic.delay.DelayManager;
-import org.terasology.logic.delay.PeriodicActionTriggeredEvent;
-import org.terasology.logic.health.event.DoDamageEvent;
+import org.terasology.logic.characters.AffectJumpForceEvent;
+import org.terasology.logic.characters.CharacterMovementComponent;
+import org.terasology.logic.characters.GetMaxSpeedEvent;
+import org.terasology.logic.characters.MovementMode;
 import org.terasology.logic.players.PlayerCharacterComponent;
 import org.terasology.physics.events.MovedEvent;
-import org.terasology.registry.In;
-import org.terasology.utilities.random.FastRandom;
-import org.terasology.utilities.random.Random;
 
 @RegisterSystem(value = RegisterMode.AUTHORITY)
 public class HypothermiaSystem extends BaseComponentSystem {
-    public static final String FROSTBITE_DAMAGE_ACTION_ID = "Frostbite Damage";
-
-    @In
-    private DelayManager delayManager;
-    @In
-    private PrefabManager prefabManager;
-    @In
-    private Context context;
-
     private static final float thresholdHeight = 60f;
-    private static final int healthDecreaseInterval = 20000;
-    private static final int initialDelay = 5000;
-    private static final int healthDecreaseAmount = 15;
     private float walkSpeedMultiplier = 0.6f;
     private float jumpSpeedMultiplier = 0.7f;
-    private Random random = new FastRandom();
 
     @ReceiveEvent(components = {PlayerCharacterComponent.class, CharacterMovementComponent.class})
     public void observeDangerZone(MovedEvent event, EntityRef player) {
@@ -67,22 +44,7 @@ public class HypothermiaSystem extends BaseComponentSystem {
         if (height < thresholdHeight && lastHeight >= thresholdHeight) {
             if (player.hasComponent(HypothermiaComponent.class)) {
                 player.removeComponent(HypothermiaComponent.class);
-                delayManager.cancelPeriodicAction(player, FROSTBITE_DAMAGE_ACTION_ID);
             }
-        }
-    }
-
-    @ReceiveEvent(components = {HypothermiaComponent.class})
-    public void onHypothermia(OnAddedComponent event, EntityRef player) {
-        delayManager.addPeriodicAction(player, FROSTBITE_DAMAGE_ACTION_ID, initialDelay, healthDecreaseInterval);
-    }
-
-    @ReceiveEvent(components = {HypothermiaComponent.class})
-    public void onPeriodicFrostbite(PeriodicActionTriggeredEvent event, EntityRef player) {
-        if (event.getActionId().equals(FROSTBITE_DAMAGE_ACTION_ID)) {
-            applyFrostbiteDamagePlayer(player);
-            applyStunEffect(player, 1000);
-            playFrostbiteSound(player);
         }
     }
 
@@ -96,25 +58,5 @@ public class HypothermiaSystem extends BaseComponentSystem {
     @ReceiveEvent(components = {HypothermiaComponent.class})
     public void modifyJumpSpeed(AffectJumpForceEvent event, EntityRef player) {
         event.multiply(jumpSpeedMultiplier);
-    }
-
-    private void applyFrostbiteDamagePlayer(EntityRef player) {
-        Prefab frostbiteDamagePrefab = prefabManager.getPrefab("ClimateConditions:FrostbiteDamage");
-        player.send(new DoDamageEvent(healthDecreaseAmount, frostbiteDamagePrefab));
-    }
-
-    private void applyStunEffect(EntityRef player, int duration) {
-        StunAlterationEffect stunAlterationEffect = new StunAlterationEffect(context);
-        //Both the instigator and the target is the player
-        //the magnitude parameter is not used by StunAlterationEffect
-        stunAlterationEffect.applyEffect(player, player, 0, duration);
-    }
-
-    public void playFrostbiteSound(EntityRef entity) {
-        CharacterSoundComponent characterSounds = entity.getComponent(CharacterSoundComponent.class);
-        if (characterSounds.deathSounds.size() > 0) {
-            StaticSound sound = random.nextItem(characterSounds.deathSounds);
-            entity.send(new PlaySoundEvent(entity, sound, characterSounds.deathVolume));
-        }
     }
 }
